feat(GRS-2): create TableDetails component
