+++ src/components/lib/table/table-details/index.ts
@@ -1,0 +1,1 @@
+export { TableDetails } from "./table-details";
