+++ src/components/lib/table/table.tsx
@@ -10,11 +10,10 @@ import {
   TablePagination,
   Paper,
   Box,
-  Typography,
-  Drawer,
 } from "@mui/material";
 import { visuallyHidden } from "@mui/utils";
 
+import { TableDetails } from "./table-details";
 import { ConditionalRender } from "../conditional-render";
 
 import type { MouseEvent, ChangeEvent, FC } from "react";
@@ -175,32 +174,7 @@ export const Table: FC<TableProps> = ({
           onRowsPerPageChange={onChangeRowsPerPage}
         />
       </Box>
-      <Drawer
-        anchor="right"
-        open={drawerOpen}
-        onClose={() => setDrawerOpen(false)}
-        sx={{ width: 400, flexShrink: 0, "& .MuiDrawer-paper": { width: 400 } }}
-      >
-        <Box sx={{ padding: 2 }}>
-          <ConditionalRender conditions={[!!selectedRepo]}>
-            <>
-              <Typography variant="h6">{selectedRepo?.name}</Typography>
-              <Typography>Description: {selectedRepo?.description}</Typography>
-              <Typography>License: {selectedRepo?.license.name}</Typography>
-              <Typography>Language: {selectedRepo?.language}</Typography>
-              <Typography>Forks: {selectedRepo?.forks_count}</Typography>
-              <Typography>Stars: {selectedRepo?.stargazers_count}</Typography>
-              <Typography>
-                Last Updated:
-                <ConditionalRender conditions={[!!selectedRepo?.updated_at]}>
-                  {new Date(selectedRepo?.updated_at || "").toLocaleDateString()}
-                </ConditionalRender>
-                <ConditionalRender conditions={[!selectedRepo?.updated_at]}>N/A</ConditionalRender>
-              </Typography>
-            </>
-          </ConditionalRender>
-        </Box>
-      </Drawer>
+      <TableDetails drawerOpen={drawerOpen} selectedRepo={selectedRepo} setDrawerOpen={setDrawerOpen} />
     </Paper>
   );
 };
