+++ src/components/views/repositories-page/index.ts
@@ -1,0 +1,1 @@
+export { RepositoriesPage } from "./repositories-page";
