+++ src/main.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 
-import { store } from "@store/store";
-
 import { App } from "./components/app";
+
+import { store } from "@store/store";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
